requirements.txt
aiofiles==23.1.0
anyio==3.7.1
fastapi==0.100.1
gcsfs==2023.3.0
google-api-python-client==2.94.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-documentai==2.14.0
google-cloud-firestore==2.10.0
google-cloud-secret-manager==2.16.4
google-cloud-storage==2.7.0
httpx==0.24.1
numpy==1.24.2
openai==0.27.6
opencv-python==4.7.0.72
sentence-transformers==2.2.2
setuptools>=66.1.1-1
shortuuid==1.0.11
sse-starlette==1.6.0
tenacity==8.2.2
thefuzz==0.19.0
tiktoken==0.4.0
pandas==1.5.3
pydantic==1.10.11
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2022.7.1
pytest==7.4.2
scikit-learn==1.2.2
uvicorn[standard]==0.21.0
wheel==0.41.2
